api/requirements.txt
Django==1.9.4
Markdown==2.6.5
cached-property==1.3.0
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-filter==0.12.0
django-url-filter==0.2.0
djangorestframework==3.3.2
requests==2.10.0
six==1.10.0
sqlparse==0.1.19
django-allauth==0.31.0
SQLAlchemy==1.1.9